pyproject.toml
[project]
name = "blunder-fix"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.116.1",
    "fsrs>=6.3.0",
    "httpx==0.28.1",
    "python-chess==1.999",
    "uvicorn==0.35.0",
]
